fix: resolver conflictos de merge y corregir migración empresa_familia
- Resueltos conflictos de Git en GeneradorMicroretos.vue tras merge con develop
- Corregido tipo de columna familia_id de integer() a unsignedBigInteger()
  para que coincida con familias.id (BIGINT UNSIGNED) y la FK no falle al migrar

+++ database/migrations/2026_04_07_100004_normalize_empresa_familia_table.php
@@ -11,8 +11,7 @@ return new class extends Migration
         Schema::table('empresa_familia', function (Blueprint $table) {
             // Añadimos familia_id FK que convive con la columna 'familia' (string) hasta backfill
             if (!Schema::hasColumn('empresa_familia', 'familia_id')) {
-                // familias.id es INT firmado (creado fuera de Laravel), hay que coincidir el tipo
-                $table->integer('familia_id')->nullable()->after('empresa_id');
+                $table->unsignedBigInteger('familia_id')->nullable()->after('empresa_id');
                 $table->foreign('familia_id')->references('id')->on('familias')->onDelete('restrict');
             }
         });
